requirements.txt
numpy==1.26.2
pandas==2.1.2
scikit-learn==1.3.2
lightgbm==4.1.0
matplotlib==3.8.1
seaborn==0.13.0
pyyaml==6.0.1
imbalanced-learn==0.11.0
jupyterlab==4.0.7
ruff==0.1.6
pytest==7.4.3

